split out constants
