+++ components/ui/Hero/characters.js
@@ -1,0 +1,46 @@
+export const COLUMNS = 120
+
+export const ROWS = 24
+
+export const CHARACTERS = [
+  '.',
+  ',',
+  ';',
+  ':',
+  '+',
+  '*',
+  '?',
+  'S',
+  '%',
+  '#',
+  '@',
+]
+
+export const CHARACTER_SEQUENCE = `
+........................................................................................................................
+........................................................................................................................
+........................................................................................................................
+........................................................................................................................
+...............,,,.................................................................................,,,..................
+...............*##;................................................................................S#S..................
+............;%%#@@S??,..........................................................................*%%@@@??*...............
+.........,;;*SS@@@@@@:.............................,;;,......................................:;;%S#@@@@@S...............
+.........;@@%..?@@@@@:.............................+@@*......................................?@@;.,#@@@@S...............
+.........,::,..?@@@@@:..............?#######%......,::?##;....,S#%............%#######?......,::,.,#@@@@S...............
+...............?@@@@@:...........:**???@@@@@@?*+...:**S@@;..**?@@@**;......;**???@@@@@#**;........,#@@@@S...............
+...............?@@@@@:........,::?@@+..%##@@@##%...*@@@@@;.,S##@@@@@%...,::S@@:..S##@@@##*........,#@@@@S...............
+...............?@@@@@:........;@@@@@+....,@@#,,....+@@@@@;..,,:@@@@@%...*@@@@@:....:@@S,,.........,#@@@@S...............
+...............?@@@@@:........+@@@@@+..?SS+;;......+@@@@@;.....;;+@@%...*@@@@@:.,%S%;;:...........,#@@@@S...............
+...............?@@@@@:........+@@@@@+..?%%,........*@@@@@?++,.,++*%%+...*@@@@@:.,?%?..............,#@@@@S...............
+...............?@@@@@:..,,,...+@@@@@*,,......,,,...+@@@@@@@@;,:#@S......*@@@@@;,,.....,,,,........,#@@@@S..,,,,.........
+...............?@@@@@:.,#@S...+@@@@@@@@:....,@@%...,,,%@@@@@@@#:,,......?@@@@@@@#,....;@@?........,#@@@@S..+@@*.........
+...............:++#@@S%%++;...:++%@@@@@S%%%%%++;......:++#@@*++.........:++#@@@@@%%%%%?++:.........++*@@@%%?++:.........
+..................*SSSS%.........+SSSSSSSSSS%............?SS,..............*SSSSSSSSSS?..............,%SSSS*............
+........................................................................................................................
+........................................................................................................................
+........................................................................................................................
+........................................................................................................................
+........................................................................................................................
+`
+  .replace(/\n/g, '')
+  .split('')
