+++ components/ui/Hero/colors.js
@@ -1,0 +1,141 @@
+import { CHARACTER_SEQUENCE, COLUMNS, ROWS } from './characters'
+
+export const COLORS = ['vibrantRed', 'vibrantBlue', 'vibrantGreen', 'trueWhite']
+
+const BLUE_CHARACTERS = [
+  [],
+  [],
+  [],
+  [],
+  [],
+  [101, 102, 103, 104, 105, 106],
+  [104, 105, 106],
+  [],
+  [22, 23, 24, 25, 97, 98, 99],
+  [47, 56, 57, 58, 59, 65, 66, 67, 68, 87, 88, 89, 105, 106, 107],
+  [22, 23, 43, 44, 45, 46, 58, 59, 60, 61, 69, 70, 71, 90, 91, 92, 105, 106],
+  [
+    40, 41, 42, 43, 56, 57, 58, 59, 70, 71, 81, 82, 83, 84, 89, 90, 91, 92, 93,
+    94,
+  ],
+  [46, 47, 48, 49, 58, 59, 60, 68, 69, 70, 71, 100, 101, 102, 103, 104],
+  [23, 24, 25, 26, 27, 45, 46, 47, 48, 56, 57, 58, 59, 85, 86, 87, 88],
+  [
+    22, 23, 24, 25, 26, 36, 37, 38, 39, 40, 41, 59, 60, 61, 62, 63, 69, 70, 71,
+    105, 106,
+  ],
+  [21, 22, 23, 37, 38, 39, 40, 41, 62, 63, 64],
+  [
+    24, 25, 26, 27, 46, 47, 48, 63, 64, 65, 67, 89, 90, 91, 104, 105, 106, 107,
+    108, 109,
+  ],
+  [
+    27, 28, 29, 30, 44, 45, 46, 47, 63, 64, 65, 66, 85, 86, 87, 88, 108, 109,
+    110, 111,
+  ],
+  [
+    25, 26, 27, 28, 29, 30, 31, 32, 33, 46, 47, 48, 49, 50, 61, 62, 63, 64, 65,
+    87, 88, 108, 109,
+  ],
+  [],
+  [],
+  [],
+  [],
+  [],
+]
+
+const WHITE_CHARACTERS = [
+  [],
+  [],
+  [],
+  [],
+  [100, 101],
+  [],
+  [20],
+  [14],
+  [],
+  [15, 55, 100],
+  [],
+  [77, 78],
+  [33, 54, 67, 85, 86],
+  [37, 38, 39],
+  [16, 17, 18, 68, 78],
+  [54, 55, 56, 105, 106],
+  [68, 81, 82, 111],
+  [20, 21, 48, 49],
+  [37, 38],
+  [],
+  [],
+  [],
+  [],
+  [],
+]
+
+const GREEN_CHARACTERS = [
+  [],
+  [],
+  [],
+  [],
+  [16, 17, 18],
+  [16, 17, 18, 19],
+  [13, 14, 15, 16, 21, 22, 99, 100, 101, 102, 103],
+  [10, 11, 12, 13, 15, 16, 17, 18, 100, 101, 102, 103],
+  [10, 11, 12, 13, 16, 17, 18, 19, 20, 21, 101, 102, 103, 104],
+  [
+    10, 11, 12, 13, 18, 19, 20, 21, 22, 41, 42, 43, 44, 45, 80, 81, 82, 83, 84,
+    85, 86, 102, 103, 104,
+  ],
+  [
+    16, 17, 36, 37, 38, 39, 40, 41, 42, 54, 55, 56, 57, 62, 63, 64, 65, 66, 67,
+    68, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 100, 101, 102, 103,
+  ],
+  [
+    16, 17, 18, 19, 20, 32, 33, 34, 35, 36, 44, 45, 46, 47, 48, 54, 55, 66, 67,
+    68, 69, 85, 86, 87, 88, 102, 103, 104, 105,
+  ],
+  [16, 17, 18, 34, 35, 36, 37, 43, 44, 45, 55, 56, 57, 74, 75, 76, 87],
+  [
+    16, 17, 18, 19, 20, 21, 22, 33, 34, 35, 36, 42, 43, 65, 66, 67, 68, 69, 74,
+    75, 76, 77, 100, 101, 102, 103,
+  ],
+  [34, 35, 55, 56, 57, 58, 66, 67, 75, 76, 77, 102, 103, 104],
+  [
+    20, 33, 34, 35, 36, 57, 58, 59, 60, 61, 72, 73, 74, 75, 76, 77, 78, 100,
+    101, 102, 103, 104,
+  ],
+  [
+    34, 35, 36, 37, 38, 39, 40, 57, 58, 59, 60, 61, 62, 76, 77, 78, 79, 80, 87,
+    88, 100, 101, 102, 103,
+  ],
+  [
+    22, 23, 24, 25, 26, 35, 36, 37, 38, 39, 40, 41, 42, 43, 58, 59, 60, 61, 62,
+    76, 77, 78, 79, 80, 81, 82, 83, 84, 104, 105, 106, 107,
+  ],
+  [
+    22, 23, 24, 41, 42, 43, 44, 45, 60, 79, 80, 81, 82, 83, 84, 85, 86, 103,
+    104, 105, 106, 107,
+  ],
+  [],
+  [],
+  [],
+  [],
+  [],
+]
+
+const colorSequence = [...CHARACTER_SEQUENCE].fill('vibrantRed')
+
+for (let rowIndex = 0; rowIndex < ROWS; rowIndex += 1) {
+  GREEN_CHARACTERS[rowIndex].forEach(colIndex => {
+    colorSequence[rowIndex * COLUMNS + colIndex] = 'vibrantGreen'
+  })
+
+  BLUE_CHARACTERS[rowIndex].forEach(colIndex => {
+    colorSequence[rowIndex * COLUMNS + colIndex] = 'vibrantBlue'
+  })
+
+  WHITE_CHARACTERS[rowIndex].forEach(colIndex => {
+    colorSequence[rowIndex * COLUMNS + colIndex] = 'trueWhite'
+  })
+}
+
+export const COLOR_SEQUENCE = colorSequence
