+++ components/ui/Hero/glitchy-ness.js
@@ -1,0 +1,13 @@
+export const DEGRADE_FACTOR = 0.6182
+
+export const FLASH_IMPACT_THRESHOLD = 0.7
+
+export const FLASH_THRESHOLD = 0.7
+
+export const GLITCH_BOUNCE_MAXIMUM_IMPACT_THRESHOLD = 0.05 // 0 - 1
+
+export const GLITCH_BOUNCE_THRESHOLD = 0.95 // 0 - 1
+
+export const GLITCH_SPEED = 100 // ms
+
+export const START_DIFFERENTIAL = 0.95
